refactor: mv 'StatementPrinter' to application layer - add reference to tests
